feat: layout in auth module was added and LoginPage was refactored.
