+++ src/auth/layout/AuthLayout.tsx
@@ -1,0 +1,61 @@
+import React, { ReactNode } from "react";
+import {
+  Card,
+  CardContent,
+  Grid,
+  Typography,
+  Box,
+  styled,
+} from "@mui/material";
+import { useTranslation } from "react-i18next";
+
+interface Props {
+  children: ReactNode;
+  title: string;
+}
+
+export const AuthLayout = ({ children, title }: Props) => {
+  // translation
+  const { t } = useTranslation();
+
+  // card style
+  const StyledCard = styled(Card)({
+    margin: "auto",
+    padding: 3,
+    borderRadius: "10px",
+    boxShadow: "0 4px 10px rgba(0,0,0,0.1)",
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    justifyItems: "center",
+    width: { md: 450 },
+  });
+
+  return (
+    <Grid
+      container
+      direction="column"
+      alignItems="center"
+      justifyItems="center"
+      sx={{ backgroundColor: "primary.main", minHeight: "100vh" }}
+    >
+      <StyledCard>
+        <CardContent>
+          <Box
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+              mb: 3,
+            }}
+          >
+            <Typography variant="h5" component="h2" sx={{ mb: 1 }}>
+              {t(title)}
+            </Typography>
+          </Box>
+          {children}
+        </CardContent>
+      </StyledCard>
+    </Grid>
+  );
+};
